리팩토링/Cell을 추상 클래스로 만들고 자식 클래스로 EmptyCell, LandMineCell, NumberCell를 생성함
