+++ src/main/java/cleancode/minesweeper/tobe/cell/EmptyCell.java
@@ -1,0 +1,28 @@
+package cleancode.minesweeper.tobe.cell;
+
+public class EmptyCell extends Cell {
+
+    private static final String EMPTY_SIGN = "■";
+
+    @Override
+    public boolean isLandMine() {
+        return false;
+    }
+
+    @Override
+    public boolean hasLandMineCount() {
+        return false;
+    }
+
+    @Override
+    public String getSign() {
+        if (isOpened()) {
+            return EMPTY_SIGN;
+        }
+        if (isFlagged) {
+            return FLAG_SIGN;
+        }
+        return UNCHECKED_SIGN;
+    }
+
+}
