+++ src/main/java/cleancode/minesweeper/tobe/cell/NumberCell.java
@@ -1,0 +1,31 @@
+package cleancode.minesweeper.tobe.cell;
+
+public class NumberCell extends Cell {
+
+    private int nearbyLandMineCount;
+
+    public NumberCell(int nearbyLandMineCount) {
+        this.nearbyLandMineCount = nearbyLandMineCount;
+    }
+
+    @Override
+    public boolean isLandMine() {
+        return false;
+    }
+
+    @Override
+    public boolean hasLandMineCount() {
+        return true;
+    }
+
+    @Override
+    public String getSign() {
+        if (isOpened()) {
+            return String.valueOf(nearbyLandMineCount);
+        }
+        if (isFlagged) {
+            return FLAG_SIGN;
+        }
+        return UNCHECKED_SIGN;
+    }
+}
